Further mem optimization - only one union-find.
This builds off of the observation that connectedness is a true relationship in the mathematical sense. If any element in the connected component is connected to the top, then so is the root. If any element in the connected component is connected to the bottom, then so is the root. Using this observation, we can efficiently track the 'connectedness' types of the components, and we don't need another Union-Find object to do it.  All tests pass, but memory is still a bit too high to pass the optional bonus.