More crazy bitwise logic to FINALLY pass mem test
This is NOT clean or pretty code.  It is NOT readable.  But it DOES finally pass the restriction that memory total out to less than 11 N^2 + 128 N + 1024.  The Weighted UF implementation has two arrays of size N^2, which approximates to 8 N^2 by itself -- adding even a single array of N^2 ints brings that up to 12 N^2.  So I used the trickery I used in a previous commit to have single ints store information about multiple grid elements.This time, there are three consecutive bits that
store data about a single element.  For ease of reasoning, I decided to store 10 elements (totaling 30 bits) in each int.  There's a small amount of wasted space, but the logic is a bit simpler this way.

Because this code is embarrassingly complicated, I may revert back to a previous stage -- but it's satisfying to finally beat that optional test.